fix: fail on consecutive EXPORT
git-annex has a bug[1] that affects special remotes that support exporttree:
Sometimes the EXPORT command is sent to a different process of the external
remote than the following TRANSFEREXPORT command and as a result, one process
receives two EXPORTs in a row.

This commit adds a check of consecutive EXPORTs so that the remote replies with
an ERROR in this case.

[1] http://git-annex.branchable.com/bugs/external_remote_export_sent_to_wrong_process